requirements.txt
fastapi
uvicorn
openai
llama-parse
faiss-cpu
python-multipart
python-dotenv
llama-index
llama-index-llms-openai
llama-index-embeddings-openai
python-docx
